✨ Add selector to change between views
